micro free

+++ includes/rt.h
@@ -253,7 +253,7 @@ int					threaddraw(void *e);
 void				mouse_press(SDL_Event *e, int press, t_envgui *env);
 Uint32				rgbafog(double dist);
 void				initfog(t_envgui *env);
-int					getmult(int aa);
+uint16_t			getmult(int aa);
 int					threadsavepic(void *adr);
 char				*ft_strnew(size_t size);
 char				*strsub(char const *s, unsigned int start, size_t len);
